fix(notifications): init local state after data fetch
