+++ src/pages/profile/ui/notifications-page/NotificationsPage.tsx
@@ -1,106 +1,18 @@
 'use client';
 
-import { useReducer } from 'react';
-import { CardSection, FloatingSaveBar, OptionGroup, Switch } from 'shared/ui';
-import { UserQueries } from 'entities/user';
-import { useUpdateNotifications } from '../../api/useUpdateNotifications';
-import { useQuery } from '@tanstack/react-query';
-import { notificationItems } from '../../config/notifications';
-import { NotificationChannel, Notifications } from '../../model/notifications';
+import dynamic from 'next/dynamic';
+import { NotificationsPageFallback } from './NotificationsPageFallback';
 
-type NotificationsState = Notifications | null;
-
-type NotificationsAction = {
-  type: 'set';
-  payload: NotificationsState;
-};
-
-const notificationsReducer = (
-  _state: NotificationsState,
-  action: NotificationsAction
-): NotificationsState => {
-  switch (action.type) {
-    case 'set':
-      return action.payload;
-    default:
-      return _state;
+const NotificationsPageContent = dynamic(
+  () => import('./NotificationsPageContent').then((mod) => mod.NotificationsPageContent),
+  {
+    ssr: false,
+    loading: () => <NotificationsPageFallback />,
   }
-};
+);
 
 function NotificationsPage() {
-  const query = useQuery(UserQueries.getMe());
-  const notifications = query.data?.notifications;
-  const [localNotifications, dispatchLocalNotifications] = useReducer(
-    notificationsReducer,
-    notifications ?? null
-  );
-  const dirty = JSON.stringify(notifications) !== JSON.stringify(localNotifications);
-  const sendSettings = useUpdateNotifications();
-
-  const handleToggle = <TChannel extends NotificationChannel>(
-    channel: TChannel,
-    key: keyof Notifications[TChannel],
-    checked: boolean
-  ) => {
-    if (!localNotifications) {
-      return;
-    }
-
-    const nextNotifications: Notifications = {
-      ...localNotifications,
-      [channel]: {
-        ...localNotifications[channel],
-        [key]: checked,
-      },
-    } as Notifications;
-
-    dispatchLocalNotifications({ type: 'set', payload: nextNotifications });
-  };
-
-  return (
-    <>
-      <CardSection
-        title="Уведомления"
-        description="Настройки почтовых и push-уведомлений."
-        className="space-y-5"
-      >
-        {Object.entries(notificationItems).map(([channel, items]) => {
-          const ch = channel as NotificationChannel;
-          return (
-            <OptionGroup
-              key={channel}
-              name={channel}
-              items={items.map((item) => {
-                const key = item.key as keyof Notifications[typeof ch];
-
-                return {
-                  key: item.key,
-                  label: item.label,
-                  input: (props) => (
-                    <Switch
-                      checked={localNotifications?.[ch][key] ?? false}
-                      disabled={query.isLoading}
-                      onCheckedChange={(checked) => handleToggle(ch, key, checked)}
-                      aria-label={item.ariaLabel}
-                      {...props}
-                    />
-                  ),
-                };
-              })}
-            />
-          );
-        })}
-      </CardSection>
-      <FloatingSaveBar
-        visible={dirty}
-        onSave={() => localNotifications && sendSettings.mutate(localNotifications)}
-        onDiscard={() =>
-          notifications && dispatchLocalNotifications({ type: 'set', payload: notifications })
-        }
-        pending={sendSettings.isPending}
-      />
-    </>
-  );
+  return <NotificationsPageContent />;
 }
 
 export { NotificationsPage };
